requirements.txt
h5py
numpy
pandas
tqdm
scikit-learn
faiss-cpu
networkx
pyarrow
neo4j
torch
pyvis
